FeatureDb-class {GenomicFeatures}	R Documentation

FeatureDb objects

Description

The FeatureDb class is a generic container for storing genomic locations of an arbitrary type of genomic features.

See ?TxDb for a container for storing transcript annotations.

See ?makeFeatureDbFromUCSC for a convenient way to make FeatureDb objects from BioMart online resources.

Methods

In the code snippets below, x is a FeatureDb object.

metadata(x): Return x's metadata in a data frame.

Author(s)

Marc Carlson

See Also

The TxDb class for storing transcript annotations.

makeFeatureDbFromUCSC for a convenient way to make a FeatureDb object from UCSC online resources.

saveDb and loadDb for saving and loading the database content of a FeatureDb object.

features for how to extract genomic features from a FeatureDb object.

Examples

fdb_file <- system.file("extdata", "FeatureDb.sqlite",
                        package="GenomicFeatures")
fdb <- loadDb(fdb_file)
fdb

[Package GenomicFeatures version 1.42.3 Index]