disjointExons {GenomicFeatures}	R Documentation

Extract non-overlapping exon parts from an object

Description

disjointExons extracts the non-overlapping exon parts from a TxDb object or any other supported object.

WARNING: disjointExons is superseded by exonicParts and will be deprecated soon. Please use improved exonicParts instead.

Usage

disjointExons(x, ...)

## S4 method for signature 'TxDb'
disjointExons(x, aggregateGenes=FALSE, 
                 includeTranscripts=TRUE, ...) 

Arguments

x
A TxDb object or any other supported object.

...
Arguments to be passed to methods.

aggregateGenes
For disjointExons : A logical. When FALSE (default) exon fragments that overlap multiple genes are dropped. When TRUE, all fragments are kept and the gene_id metadata column includes all gene ids that overlap the exon fragment.

includeTranscripts
For disjointExons : A logical. When TRUE (default) a tx_name metadata column is included that lists all transcript names that overlap the exon fragment.

Details

disjointExons creates a GRanges of non-overlapping exon parts with metadata columns of gene_id and exonic_part. Exon parts that overlap more than 1 gene can be dropped with aggregateGenes=FALSE. When includeTranscripts=TRUE a tx_name metadata column is included that lists all transcript names that overlap the exon fragment. This function replaces prepareAnnotationForDEXSeq in the DEXSeq package.

Value

A GRanges object.

Author(s)

disjointExons was originally implemented by Mike Love and Alejandro Reyes and then moved (and adapted) to GenomicFeatures by Valerie Obenchain.

See Also

exonicParts for an improved version of disjointExons.

Examples

## TODO

[Package GenomicFeatures version 1.42.3 Index]