getPromoterSeq {GenomicFeatures}	R Documentation

Get gene promoter sequences

Description

Extract sequences for the genes or transcripts specified in the query (aGRanges or GRangesList object) from a BSgenome object or an FaFile.

Usage

  ## S4 method for signature 'GRangesList'
getPromoterSeq(query, subject, upstream=2000, downstream=200, ...)
  ## S4 method for signature 'GRangesList'
getPromoterSeq(query, subject, upstream=2000, downstream=200, ...)
  ## S4 method for signature 'GRanges'
getPromoterSeq(query, subject, upstream=2000, downstream=200, ...)

Arguments

query
A GRanges or GRangesList object containing genes grouped by transcript.

subject
A BSgenome object or a FaFile from which the sequences will be taken.

upstream
The number of DNA bases to include upstream of the TSS (transcription start site)

downstream
The number of DNA bases to include downstream of the TSS (transcription start site)

...
Additional arguments

Details

getPromoterSeq is an overloaded method dispatching on query, which is either a GRanges or a GRangesList. It is a wrapper for the promoters and getSeq functions. The purpose is to allow sequence extraction from either a BSgenome or FaFile.

Default values for upstream and downstream were chosen based on our current understanding of gene regulation. On average, promoter regions in the mammalian genome are 5000 bp upstream and downstream of the transcription start site.

Value

A DNAStringSet or DNAStringSetList instance corresponding to the GRanges or GRangesList supplied in the query.

Author(s)

Paul Shannon

See Also

intra-range-methods ## promoters method for IntegerRanges objects intra-range-methods ## promoters method for GenomicRanges objects getSeq

Examples

  library(TxDb.Hsapiens.UCSC.hg19.knownGene)
  library(BSgenome.Hsapiens.UCSC.hg19)
 
  e2f3 <- "1871"  # entrez geneID for a cell cycle control transcription
                  # factor, chr6 on the plus strand

  transcriptCoordsByGene.GRangesList <-
     transcriptsBy (TxDb.Hsapiens.UCSC.hg19.knownGene, by = "gene") [e2f3]
    # a GrangesList of length one, describing three transcripts

  promoter.seqs <- getPromoterSeq (transcriptCoordsByGene.GRangesList,
                                   Hsapiens, upstream=10, downstream=0)
    # DNAStringSetList of length 1
    # [["1871"]] GCTTCCTGGA GCTTCCTGGA CGGAGCCAGG

[Package GenomicFeatures version 1.42.3 Index]